pyproject.toml
[project]
name = "orbit-ai"
version = "0.1.0"
description = "Personal AI assistant powered by LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=3.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic[email]",
    "beautifulsoup4",
    "wikipedia",
    "langchain",
    "langchain-community>=0.0.74",
    "langchain-experimental>=0.0.32",
    "langchain-google-genai>=0.0.7",
    "langgraph>=0.2.17",
    "playwright>=1.45.0",
    "requests>=2.31.0",
    "typing-extensions>=4.7.1",
    "markdown-pdf>=1.0.0",
    "google-auth>=2.22.0",
    "python-docx",
]

[project.scripts]
orbit-ai = "app:main"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]